requirements.txt
flask
requests
numpy==1.19.5
tensorflow==2.4.1
python-dotenv
gunicorn